LabelDroid/requirement.txt
numpy
opencv-python
Pillow
torch
torchvision
tqdm
nltk
pycocotools
gensim
tensorflow
tensorflow_hub
sklearn
flask
flask_restful